backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.6.3
elasticsearch[async]==8.13.0
aiohttp==3.9.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pydantic-settings==2.3.0
pydantic[email]==2.7.0
httpx==0.27.0
numpy==1.26.4
python-dateutil==2.9.0
